+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### X.X.X (coming soon)
 
+Other:
+  - Improve error message when write functions are called with empty objects.
+
+
 ### 10.0.0 (20 July 2021) BREAKING [read the announcement](./orga/releases/10.0.0.md)
 
 Breaking: